Add ability to traverse multigraphs
Now works with graphs invovling multiple edges from one node to
another.

This was implemented by added an id attribute to the Edge class.
The normal case is that the id will be 0, but it is incremented
for each additional edge between a given pair of nodes. This feels
like a hack but should work for now.